feat: change amount to integer to represent minor units
- Update database schema from DECIMAL to BIGINT for amount storage
- Modify domain model, DTOs, and entities to use Long type for amount
- Adjust validation logic to work with integer comparisons
- Update test data to use integer amounts representing minor units
- Simplify TransactionResponse DTO to return only essential fields
- Add provider rejection logic for specific amount (999999)
- Update README with comprehensive project documentation

+++ src/main/java/com/orchestrator/transaction/application/dto/TransactionResult.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
+
 import java.time.Instant;
 import java.util.UUID;
 
@@ -26,7 +26,7 @@ public class TransactionResult {
     @JsonProperty("client_transaction_id")
     private String clientTransactionId;
     
-    private BigDecimal amount;
+    private Long amount;
     private String currency;
     private TransactionStatus status;
